Allagan Tools v1.2.0.17 (#1227)
* Allagan Tools v1.2.0.17

This is a bug fix release. Fixed an issue when you initially add in a craft list. Have put in more code to help mitigate a potential saving crash. If anyone is crashing reliably and knows their way around a debugger, can they attach it and get a stack trace please.

* Update manifest.toml