TextToTalk v1.28.3 (#2904)
* TextToTalk v1.28.3

This update finally merges the testing version of the plugin back to the stable version of the plugin.
I'll be paying closer attention to the debug channel this update as there have been past issues with config
updates between 1.24 and 1.25 that may still show up in some cases.
Please send your plugin config if this occurs so I can quickly patch any remaining instances of the issue.

Changes since 1.28.2
- Improves logging during config migrations for easier debugging in the future

Changes since 1.24.2
- Adds support for ElevenLabs TTS
- Adds Cortana voices to the "Don't see all of your voices?" button

* Fix manifest