Update manifest.toml
I noticed my client closed recently and I am unsure if it is related to the plugin.
It could have just been a random close but I didn't receive anything in logs about the plugin or a crash error so I am unsure if it is the issue. I am sure you guys will experience it too if it is an ongoing issue.