Absolute Roleplay [Testing] (#4706)
* Absolute Roleplay

* Absolute Roleplay

* Trying to fix error

* Updated InternalName

* Update manifest.toml

* Changing Folder name appropriately

* Update manifest.toml

Figured out there was a small chance of auto login not working and actually locking the connection for the user, This fixes it.

* Update manifest.toml

Set messages for passwords not matching and scaled the restoration window correctly, as well as moved away from config values when sending data to server as it relied on the user checking remember me.

* Needed fix for registration

Fixed an issue where the registration key would some times not work after registering.

* Update manifest.toml

Few minor bug fixes.

* Update manifest.toml

Updated Imaging to nolonger download images but simply read from url

* Update manifest.toml

Changed Constants.cs to Defines.cs
Just a name change

* Update manifest.toml

Reverting to working build

* Update manifest.toml

Removed Unneeded File

* Update manifest.toml

Trying to rollback further

* Update manifest.toml

* Update manifest.toml

* Update manifest.toml

Updated Discord Link in MainPanel.cs

* Absolute Roleplay plugin update

I had to update the client and server to use ssl.
Took me a little bit but it's done now

* Version update attempt

Trying to update the version.

* Update manifest.toml

Fixed a ui issue with having textwrapped instead of safetextwrap

* Quick update

Some qauality of life changes and some minor adjustments
mainly::
properly retrieving profiles from friend list or chat.
added a way for users to change if the window closes on connect
also clearing packets every time InitializePackets is called to avoid issues.
removed Framework update function and replaced it with onlogin and ctor. Thank you MidoriKami <3

* Update manifest.toml

UI changes and a new request and response from the server.
Added tooltip functionality for when you hover your mouse over a player that is connected with you or is not set to private profile.
Overhauled the options window.

* Update manifest.toml

* Update manifest.toml

Fixed an issue with profiles not resetting on character swap. 
Added the ability to save and load your profile locally. Fixed some scaling issues with the ui and very large dalamud ui resolutions. Also added the ability to delete your profile.

* Update manifest.toml